Rename Chatbot component file and update import paths; adjust chatbot styling options
