+++ frontend/app/layout.js
@@ -2,7 +2,7 @@ import { Inter } from "next/font/google";
 import "./globals.css";
 import Navbar from "@/components/Navbar";
 import Footer from "@/components/Footer";
-import Chatbot from "@/components/Chat/Chatbot";
+import Chatbot from "@/components/Chatbot";
 import { ThemeModeScript } from "flowbite-react";
 import request from "@/lib/request";
 import { ToastContainer } from "react-toastify";
